agentic-weather-update/backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
data3-network-variables==0.0.6
